backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "mecab-python3>=1.0.12",
    "nodeenv>=1.10.0",
    "openai-whisper>=20250625",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "unidic-lite>=1.0.8",
    "uvicorn[standard]>=0.40.0",
    "yt-dlp>=2025.12.8",
]
